+++ metadata.rb
@@ -1,6 +1,6 @@
-name             'mkdirs-nativex'
-maintainer       'NativeX'
-maintainer_email 'brett.stime@nativex.com'
+name             'mkdirs-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
 description      'Creates [empty] directories configured by a node attribute list'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))